Split psa_switch.h into two files
