+++ backends/bmv2/psa_switch/psaSwitch.h
@@ -17,6 +17,7 @@ limitations under the License.
 #ifndef BACKENDS_BMV2_PSA_SWITCH_PSASWITCH_H_
 #define BACKENDS_BMV2_PSA_SWITCH_PSASWITCH_H_
 
+#include "psaProgramStructure.h"
 #include "ir/ir.h"
 #include "lib/gmputil.h"
 #include "lib/json.h"
@@ -116,45 +117,10 @@ class PsaSwitchExpressionConverter : public ExpressionConverter {
     }
 };
 
-class PsaProgramStructure : public ProgramStructure {
-    P4::ReferenceMap*    refMap;
-    P4::TypeMap*         typeMap;
-
+class PsaCodeGenerator : public PsaProgramStructure {
  public:
-    // We place scalar user metadata fields (i.e., bit<>, bool)
-    // in the scalars map.
-    ordered_map<cstring, const IR::Declaration_Variable*> scalars;
-    unsigned                            scalars_width = 0;
-    unsigned                            error_width = 32;
-    unsigned                            bool_width = 1;
-
-    // architecture related information
-    ordered_map<const IR::Node*, std::pair<gress_t, block_t>> block_type;
-
-    ordered_map<cstring, const IR::Type_Header*> header_types;
-    ordered_map<cstring, const IR::Type_Struct*> metadata_types;
-    ordered_map<cstring, const IR::Type_HeaderUnion*> header_union_types;
-    ordered_map<cstring, const IR::Declaration_Variable*> headers;
-    ordered_map<cstring, const IR::Declaration_Variable*> metadata;
-    ordered_map<cstring, const IR::Declaration_Variable*> header_stacks;
-    ordered_map<cstring, const IR::Declaration_Variable*> header_unions;
-    ordered_map<cstring, const IR::Type_Error*> errors;
-    ordered_map<cstring, const IR::Type_Enum*> enums;
-    ordered_map<cstring, const IR::P4Parser*> parsers;
-    ordered_map<cstring, const IR::P4ValueSet*> parse_vsets;
-    ordered_map<cstring, const IR::P4Control*> deparsers;
-    ordered_map<cstring, const IR::P4Control*> pipelines;
-    ordered_map<cstring, const IR::Declaration_Instance*> extern_instances;
-    ordered_map<cstring, cstring> field_aliases;
-
-    std::vector<const IR::ExternBlock*> globals;
-
- public:
-    PsaProgramStructure(P4::ReferenceMap* refMap, P4::TypeMap* typeMap)
-        : refMap(refMap), typeMap(typeMap) {
-        CHECK_NULL(refMap);
-        CHECK_NULL(typeMap);
-    }
+    PsaCodeGenerator(P4::ReferenceMap* refMap, P4::TypeMap* typeMap)
+            : PsaProgramStructure(refMap, typeMap) {}
 
     void create(ConversionContext* ctxt);
     void createStructLike(ConversionContext* ctxt, const IR::Type_StructLike* st);
@@ -168,68 +134,6 @@ class PsaProgramStructure : public ProgramStructure {
     void createDeparsers(ConversionContext* ctxt);
     void createGlobals();
     cstring convertHashAlgorithm(cstring algo);
-
-    std::set<cstring> non_pipeline_controls;
-    std::set<cstring> pipeline_controls;
-
-    bool hasVisited(const IR::Type_StructLike* st) {
-        if (auto h = st->to<IR::Type_Header>())
-            return header_types.count(h->getName());
-        else if (auto s = st->to<IR::Type_Struct>())
-            return metadata_types.count(s->getName());
-        else if (auto u = st->to<IR::Type_HeaderUnion>())
-            return header_union_types.count(u->getName());
-        return false;
-    }
-};
-
-class ParsePsaArchitecture : public Inspector {
-    PsaProgramStructure* structure;
- public:
-    explicit ParsePsaArchitecture(PsaProgramStructure* structure) :
-        structure(structure) { CHECK_NULL(structure); }
-
-    void modelError(const char* format, const IR::INode* node) {
-        ::error(ErrorType::ERR_MODEL,
-                (cstring(format) + "\nAre you using an up-to-date 'psa.p4'?").c_str(),
-                node->getNode());
-    }
-
-    bool preorder(const IR::ToplevelBlock* block) override;
-    bool preorder(const IR::PackageBlock* block) override;
-    bool preorder(const IR::ExternBlock* block) override;
-
-    profile_t init_apply(const IR::Node *root) override {
-        structure->block_type.clear();
-        structure->globals.clear();
-        return Inspector::init_apply(root);
-    }
-};
-
-class InspectPsaProgram : public Inspector {
-    P4::ReferenceMap* refMap;
-    P4::TypeMap* typeMap;
-    PsaProgramStructure *pinfo;
-
- public:
-    InspectPsaProgram(P4::ReferenceMap* refMap, P4::TypeMap* typeMap, PsaProgramStructure *pinfo)
-        : refMap(refMap), typeMap(typeMap), pinfo(pinfo) {
-        CHECK_NULL(refMap);
-        CHECK_NULL(typeMap);
-        CHECK_NULL(pinfo);
-        setName("InspectPsaProgram");
-    }
-
-    void postorder(const IR::P4Parser *p) override;
-    void postorder(const IR::P4Control* c) override;
-    void postorder(const IR::Declaration_Instance* di) override;
-
-    bool isHeaders(const IR::Type_StructLike* st);
-    void addTypesAndInstances(const IR::Type_StructLike* type, bool meta);
-    void addHeaderType(const IR::Type_StructLike *st);
-    void addHeaderInstance(const IR::Type_StructLike *st, cstring name);
-    bool preorder(const IR::Declaration_Variable* dv) override;
-    bool preorder(const IR::Parameter* parameter) override;
 };
 
 class ConvertPsaToJson : public Inspector {
@@ -238,11 +142,11 @@ class ConvertPsaToJson : public Inspector {
     P4::TypeMap *typeMap;
     const IR::ToplevelBlock *toplevel;
     JsonObjects *json;
-    PsaProgramStructure *structure;
+    PsaCodeGenerator *structure;
 
     ConvertPsaToJson(P4::ReferenceMap *refMap, P4::TypeMap *typeMap,
                      const IR::ToplevelBlock *toplevel,
-                     JsonObjects *json, PsaProgramStructure *structure)
+                     JsonObjects *json, PsaCodeGenerator *structure)
         : refMap(refMap), typeMap(typeMap), toplevel(toplevel), json(json),
           structure(structure) {
         CHECK_NULL(refMap);
